// angular-sample/angular-webapp/client/app/checklist/models/workspace.model.ts
export class Workspace {
    Author: string;
    Modifier: string;
    Name: string;
    PictureId: string;
    Id: string;
    CreatedOn: string;
    ModifiedOn: string;
    CreatedBy: string;
    ModifiedBy: string;
    IsDeleted: boolean;
    LCid: number;
    Version: string;
    isSelected: boolean;
}